Commit it all in one go with working examples
I wanted to split this into as-small-as-possible individual commits
and small PRs but have no time now, pushing as one to make it
accessible for review. maybe split into smaller if needed for
review, but it does not appear to large for one commit now.

Let's see.

It is working with the tow updated examples on my system, so
it has real-live validation.

The maybe not-obvious small changes and additiona are fixes
for the real-live query from examples/async/list-connections-async.py
by reading the many different VPNs and other connections on my system.