Genereate the integration of the remaining settings objects
The previous commit manually added the set of settings
objects which I have on my machine.

This commit generates the full integration of settings objects.

Like with the previous commit, this could be split into many
small commits, but it should be ok to review like the last was.

Like before, the small changes are fixing up e.g. imperfect
ordering of manually ordered lists.

I fixed names where I used "wifi" before to "wireless" as
this is the wording the wording from the NM settings objects.

Like before, this is tested with the two updated examples
on my system which has many connection types defined.