enums.py: Fix ConnectionType and show it's use.
ConnectionType gives a 1:1 map from a DeviceType.name to
a profile.connection.connection_type value, which can be
used to find usable connections for a given DeviceType.

Fix ConnectionType (added very recently, not yet released)
to not use ConnectionType.WIRED for an Ethernet connection
but ConnectionType.ETHERNET, which matches DeviceType.ETHERNET.

Since ConnectionType is new, the fix is safe to apply now.

Demonstrate it's use in the netdevinfo examples like shown
in PR #36 with conn_type = getattr(ConnectionType, dev_type).