tools/generate-settings-dataclasses.py: Generate attr docstrings(1/2)
Prepare to generate docstrings for the attributes of settings_classes.
Even though they are not parsed by python itself, they can be
used to generate documentation for the settings dataclasses.

To not require manual edits after code generation, it was needed to
implement sorting of required attributes like connection_id, -type
and uuid before other attributes. The new functions for this are
derived from the existing sort function for the settings_classes
where we need the connection as first attribute of the profile
because likewise, it is not an optional attribute.

The switch to enable generating the attribute docstrings is deferred
to the next commit, to make reviewing this change above (which does
not change the established code) and the minor change below easier:

There is a minor change in the established code but only regarding
sdbus_async/networkmanager/settings/bond.py whose attributes
are not actually required and should be optional. Making them
required was a quick hack around some issue I don't recall the
exact details of, and should be adressed at its source anyway.

This is the only code change as result of running the updated
tool, and the next commit will enable the switch to generate
the docstrings.

To aid in understanding and (in the future, reworking) the
code for use in the new upcoming generator based on Jinja2,
this also adds a few comments and docstrings for navigating
the code.

I guess this may help be a step towards adressing PR #31,
for documenting new Settings helpers classes.