Fix ConnectionType for "gsm" to match DeviceType.MODEM
Apply the same kind of fix like in #36 for ETHERNET where we fixed
ConnectionType.ETHERNET to match DeviceType.ETHERNET:

Use ConnectionType.MODEM to match DeviceType.MODEM for use with
getattr() like shown in #36: `getattr(ConnectionType, "MODEM")`.

Like with the rename from .WIRED to .ETHERNET in #36, this rename
from .GSM to .MODEM is ok because the Enum ConnectionType is new,
and wasn't in a release of this library yet.